Introduced Design System Helper
To encapsulate the two methods required by most controllers to toggle between design system view and legacy view.
